Instal tasks: use a dedicated variable to hold install error message. Re #66.
Reorganization in preparation for introducin a new install error on Windows 10 (provided that Windows 10 Version 22H2 is a minor update from Vibranium branch): move the install error message to before the error dialog and assign the text (tring) to a dedicated variable. This allows the install erro dialog to present whatever error message it receives.