Notepad (Windows 11): add comments on text repetition problem fix.
The issue of NVDA repeating the previous line's text when Enter is pressed is resolved in Notepad 11.2203.10.0. This build also resolves line repetition problem where there was text, followed by a blank line, then end of file. Since a Notepad document class is defined to handle text announcement problem, this will be kept for now for backward compatibility and wil be removed once the fixed Notepad release is published to stable Windows 11 build (the status bar/Notepad document detection wil be modified, however).