Settings (Windows 10): do not cut off live regions if update action button such as 'restart now' shows up.
When update installation complets, update action buttons such as 'restart now' is presented. However with live regions being cut off, button labels and related live region text such as a reminder to restart the system is not heard in its entirety. Therefore detect update action button (next to updates list and is a UIA button element) and do not cut speech if action button is showing (the update item live region is the child of the actual update item, in turn a child of the updates list, and next to the updates list is action button if present).