WinApps constructor: add additional god UIA window class names. Re #72.
NVDA Core's good UIA window class names is a tuple, and the add-on version is a list. Therefore first convert the core version to a list, attach the add-on version of good UIA window class names list, then convert the whle thing to a tuple and assign it to the NVDA Core version. Throughout this process, a debug log will inform users about additional good UIA window class names.