Calculator: remove graphing calculator fix to align code with NVDA Core.
As of NVDA 2022.2 alpha (late April 2022), the omly difference in UIA notification event handler is graphing calculator repetition fix. While this fix is documented in the readme, it isn't included the add-on changelog. Besides, other parts of the Calculator app module are part of NVDA Core. Tehrefore as part of preparation for removing the Calculator app module (unless issues arise in 2022), remove graphing calculator fix to align all parts of the app module with that of NVDA Core.