WinApps/File Explorer good UIA windows: add touch keyboard button window to a list of good UIA window class names.
It appears build 22616's reversal of system tray changes from earlier Insider preview builds is not allowing touch keyboard button to be navigable with mouse and/or touch. Thankfully there is a dedicated window class name for this button, therefore add this class name (TIPBand) to the list of good UIA window class names for File Explorer/Windows 11.