WinApps/Windows 11 UI: edit the description of Shell_TrayWnd class.
It turns out Shel-TrayWnd class is the root element of Windows 11 shell UI elements in addition to housing Start, Widgets and other buttons (evidenced by UIA tree). Therefore edit the description of this class window as it might be possible to simplify the app module patching method by checking for this class and a limited number of other windows somehoow.