WinApps/File Explorer: remove File Explorer top-level window class name for This PC and friends. Re #72.
Remove CabinetWClass (This PC and other top-level windows) as it causes regression where NVDA says 'system' when Alt+Space is pressed. While this removes the ability to explore File Explorer command bar items with mouse and/or touh, removing Alt+Space announcement is a bigger reason for removing this class name.