Settings: allow name change event for update announcement to work with selective UIA event registration on.
Even if name change event is fired, with selective UIA event registration on, it is not caught by NVDA as this is beyond the scope of the currently focused control. Fortunately, wen update entries appear, a ist is shown and can be focused, and focus entered event must be defined to detect items in this list (as items are inside a container control). Once update entries are located, look for status text (entry.firstChild.lastChild) and add it to local event handler group so update status can be announced as if selecive event registration is off. Note that this will work on Windows 11 22H2 and later due to focus movement problem in Windows 11 21H2. Also, in order for this workarond to work correctly, users must focus on updates list as soon as it appears after checking for updates as the workaround is tied to focus entered event. While quite hacky, this workaround allows update status to be announced while focused on Settings app as long as Windows UPdate category/updates list is showing, similar to how updates can be announced in Windows 10. Lastly, this workaround may not work properly if multiple updates are pending, but tests indicate that at least one item status is read.