Microsoft Solitaire Collection: version 4.13 and later improves accessibility by adding labels to cards (deprecate the app module).
The whole point of Solitaire app module was adding labels to cards. This was the case in version 4.12 and earlier where UIA localized control type must be fetched. Version 4.13 and later improves accessibility by adding labels to cards (at least suits and ranks but not stacks). Therefore, deprecate the entire app module and schedule it for removal in a future add-on release. Until the app module is removed, detect old version by noting that 'unknown' controls (cards) have no names.