Calculator: remove the app module.
Windows 10 (and now 11) Calculator app module was one of the first app modules created for Windows App Essentials in 2015. Removing this module is significant not only because NVDA 2022.2 includes command announcement fix from the latest code from the app module, it also marks a transition from active development to more toward a maintenance mode for the add-on as a whole. Calculator app module may return if needed, but for now, farewell, Windows 10 Calculator app module.