Winapps/item status event: detect duplicate events through eventHandler.isPendingEvents.
Sometimes item status event is fired by the same object multiple times. Thankfully, this can be detected if eventHandler.isPendingEvents says 'True'. Therefore use this result and also make sure the item status from the 'new' object is not the same a the one being handled. This chnage eliminates the need for caching item status text and runtime ID, simplifying the event handler.