WinApps/drop target: locate drop target effect text from parent objects if possible.
Sometimes (notably in Windows 10 Action Center/Windows 11 Quick settings), an element other than the focused object fires drop target events. In some cases these events are fired not by the object that supposedly reports the effects of this event (in Windows 11 quick settings, it is the grouping object for settings items that actually fires this event when in fact it is thel ist that contains action items that is supposed to be the target of drop taret events set), resulting in drop target effect text being empty. Therefore attempt to locate effect text by traversing parent objects. For the most part, this will result in finding drop target effect text somewhere in the ancestor chain, but it might be possible that this text may not be found even after looking up desktop (shell) object. With this change, it is now possible for NVDA to (finally) announce drop target effect text when rearranging action center/quick settings items (after at least three years of looking for solutions).