WinApps/drop target: traverse  focus ancestors instead of going up the parents when looking for elements with drop target effect text.
Rather than using a loop to traverse  parent elements, fetch focus ancestors which provides an accurate picture of where the focus is and may eventually lead to the actual element with drop target effect text. Note that ancestors list must be reversed for optimization purposes (often drop target event is fired by an ancestor that is above the element that was dragged).