WinApps/virual desktop switching: remove event handler's handle desktop switch attribute check.
As the pull request for NVDA Core issue 5641 is stalled (as of September 2022), do not check for 'handle desktop switch' atribute in event handler module - just assume the fix is not included in NVDA for now. Attribute check may return if the NVDA Core pull request is picked up in the future.