SearchUI: revived to provide foundations to support Search Highlights in Windows 11.
Revive SearchUI app module (removed in 22.09) so Search Highlights can be announced in Windows 11 (21H2 and 22H2). Search Highlights is a text that is displayed, offering a top search suggestion of the day. This turns out to be part of an element that is the last child of the Start search box, so SearchUI app module must b used to house fixes.