SearchUI/search field: use description getter to obtain search highlights text.
Instead of a focus event, extend NVDA Core's Start search field class by adding a description getter which returns search highlights text if the last child element is present. This allows NVDA to announce search field label correctly if highlights is shown or not while not searching and searching, respectively (mostly affects speech output). As part of this commit, reintroduce overlay classes chooser and tlel it to recognized the enhanced Start search field and move to NVDA Core version, as well as transfer gain focus event from the app module to the description getter in the search field class and edit imports.