BuildVars: remove minimum Windows version flag.
As Windows 10 and 11 uses different build ranges and since Windows 10 21H2 wil be out of suport for consumers in June 2023, it makes no sense to keep minimum version flag (it can confuse people into thinking that any build above 19044 is supported when it is not). Besides, it is install tasks that will check minimum Windows version anyway.