Instal tasks: introduce a combined Windows release erorr message. Re #78.
Introduce a new message to address both Windows 10 and feature update compatibility situations (to be shown when not using Windows 10 and/or unsupported feature update is in use). This message will be similar to the one shown for feature update message except it will add current release name and build number to help users identify the Windows release they are running. Supported releases will now include values from Windows releases dictionary (Windows 10/11/Server) and it comes from a dedicated variable. The variable name for the message is same as other error messages and is designed to be set differently based on a test flag (upcoming commit).