Install tasks: remove Windows 10 requirement comment. Re #79.
After six years, remove the comment that defined Windows App Essentials installer: Windows 10 or later requirement. This was added in 2016 to catch Windows 7 and 8.x as these releases were popular, more so for Windows 7 then. As of January 2023, Windows 7 and 8.1 will no longer be supported in any form that is familiar to consumers and businesses (extended security updates (ESU) for Windows 7, extended support for Windows 8.1). Besides, only feature updates for Windows 10 and 11 that are supported for consumers will be supported, making Windows 10 requirement comment redundant.