WinApps: do not add additional events and invoke the global event handler group if additional events and properties maps are empty. Fixes #81.
Do not proceed to adding events and properties if events and properties maps are empty, designed to avoid initializing a global event handler group that comes with the add-on. This is more so wen running on NVDA 2022.4 as the only residents of additional events and properties maps (drag and drop events and effects) are part of NVDA Core.