WinApps: remove additional event and property maps. Re #80.
Additional event and property maps were used to house events not recognized by NVDA Core. This was useful in that the add-on can enhance UIA support at the cost of being seen as a hack (events and properites in these maps were forcefully added to UIA handler's event and property maps, later asking UIA handler to handle them, sometimes at the global level). In addition to adding support for controller for event, the key part of these maps was drag and drop support, later adding layout invalidated and other events. Now that iproved support for accessible UIA drag and drop is part of NVDA Core, this clears event and property maps, therefore remove them (for now).