Modern keyboard: define Windows 11 IME candidate item object to ignore UIA element selected event.
Becasue focus event is fired on IME candidate items (including when selecting entries for hardware keyboard suggestions), there is really no need to handle elemenet selected event (if handled, this will result in double speaking). Therefore, define a custom overlay class specific to Windows 11 modern keyboard to ignore UIA element selected event, named the same as the built-in NVDA Core class for compatibiilty with the rest of the app module.