Notepad 11: repositoin status bar index to support newer Notepad releases.
Notepad 11.2212 introduces tabbed Notepad interface. This also introduces a change to UIA tree for Notepad in that status bar is no longer the eighth position in the UIA tree. Thankfully, status bar is the second to last item in the UIA tree across Notepad 11 releases (icluding tabbed Notepad), so reposition status bar index.