Explorer (app module): add NVDA Core version of File Explorer app module.
Bundling File Explorer ap module with the add-on again: in 2021 a discussion took place in NVDA add-ons list about how to deal with a stack of app modules with the same executable name. A global plugin based solution was proposed and was implemented in Explorer Enhancements add-on where the global plugin will add name change and focus entered events for announcing empty folder text. Thankfully File Explorer enhancements in Windows App Essentials do not override these events, so it can be used safely with other add-ons.