WinApps/File Explorer: define a dedicated (and private) empty folder detector to centralize empty folder annoncement and element lookup.
There is an issue with focus entered event: if there were items in the folder with items list being focused, 'this folder is empty' will not be announced if items are deleted and focus entered event is fired on the items list. In this situation, last child will not work becasue NVDA will find itself stuck in a nonexist list item (no child, no sibling). The only way out is using obj.parent. Coupled with the fact that name change event will locate the empty folder text anyway, take this opportunity to define a private 'empty folder detector' that will anounce empty folder text if this element is found. The bulk of this new private detector comes from name change event with UIA tree traversal - the only remnant of Explorer Enhancements code is empty folder announcement. To avoid unnecessary foregrond lookup, take in the object the event is for as a parameter i.e. from name change event, pass in title element.