Explorer/taskbar items: catch value error if UIA tree does not provide a complete list of chlidren.
Sometimes an XAML control sits in the middle of the taskbar, causing UIA tree to not show the complete list of children, leading to value error when fetching position info. Therefore see if the current taskbar item is part of the taskbar items list, and if not, return early. This is mostly seen when using input devices other than the keyboard.