Settings/taskbar/search combo box: do not add label if the combo box's parent is search grouping item.
Taskbar/search combo box label issue is resolved through a search grouping. Therefore detect this grouping (parent of the combo box element) and do not add search label accordingly. This is the case with build 22621.1483 (release preview), with the workaround to be removed once this fix comes to general availability channel.