File Explorer/this folder is empty: remove the feature entirely due to a critical bug that cannot be resolved easily.
Consider the following: an app that does not raise system focus event when switched to is running, user copies app title, opens File Explorer, creates a new folder giving it the same name as the first app title, then opens the newly created folder. When switching betwen apps, NVDA may announce the same thing when switching between File Explorer and the first app. This is caused by the solution found in the add-on: the add-on relies on UIA element selected and name change events to look for 'this folder is empty' text, not system focus event becasue focus event is not raised in UIA. As this is indeed a bug with File Explorer's UIA implementation, coupled with the critical bug described, 'this folder is empty' announcement feature is hereby removed.