Install tasks: only present Windows 11 releases if 19045 < build < 22000.
Partial return of adaptive versions list: if running on a build between Windows 1 22H2 and 11 21H2 (say, Server 2022/20348), present only Windows 11 builds as downgrading requires a clean install or an equivalent (removing apps and settings). Previously, all supported releases were shown, but with end of support for Server 2022 from the add-on, there is a build range gap, so it makes more sense to display Windows 11 releases on 20000 and 21000 build ranges.