Readme: add the concept of emulated fixes.
In some cases, NVDA (through the add-on) may emulate fixes included in Insider Preview builds which might be removed in future preview builds (one example is rearranged taskbar item announcement introduced in build 25267 and removed in 25352). Therefore note these changes and declare that emulated fixes may be removed in future add-on releases.