WinApps/dialog detection: resolved in NVDA 2024.1.
Turns out there was a flaw with IsDialog property check in NVDA Core: if IsDialog property is False, NVDA would not proceed to checking dialog class names, causing dialogs such as Windows Update restart dialog (from Insider Preview builds) to not be announced (dialog is detected but the content is not reported). This issue (a regression really) is resolved in NVDA 2024.1.