Calculator: app module returns to resolve NVDA Core isseu 15923.
NVDA issue 15923/Microsoft Calculator issue 2097: if an operator is surrounded with identical operand (for example, 2+2), UIA notification is not fired when result command (Enter, for example) is pressd. Therefore, force no results to be False so results can be announced. This should be a temporary workaroudn and should be removed when resolved by either NVDA or Calculator (the latter is preferred).