Modern keyboard/window open event internal: simplify dictaiton/voice typing autoamtion Id detection and handling.
Reduce conditionals level by one - rather than checking for AutomationId's and doing nested if statements, just check the string in one go. In other words:
* Old method: a main if statement to check a list of Automation Id's, followed by if/else for dicttion/voice typing Automation Id's.
* New method: if/elif, with each line testing a different Automation Id.
The old method was defined back in the day when it was  desirable to run a single while loop to add dictation/voice typing elements in one go, but with list concatention operation, this is no longer the case (while loop was removed). Therefore, if/elif statements are sufficient.