Install tasks/error message: add a dedicated error message for 32-bit Windows 10 detection.
Present a message when trying to install the add-on on 32-bit Windows 10 systems. The wording wil reference Windows 10 (it is highly unlikely that Windows 8.1 32-bit users wil attempt to install the add-on). A flag wil be passed ito message presentation function to show either the generic 'unsupporte Windows release' message or the 32-bit specific message.