Modern keyboard: hide window open and notification events if NVDA 2024.2 or later is running.
Because contents from UIA window open and notification events or workaronds are now part of NVDA, there is no need to for the add-on itself to handle them. Therefore, pass these events to nVDA Core directly, and one way to do so is hiding these event handlers if NVDA 2024.2 or later is running.