Modern keyboard/UIA element selected event: check navigator objects to see if a different part of emoji panel is showing.
Consider the following: a Windows 11 user opens emoji panel (Windows+Period), press Shift+Tab twice to move to categories list, then selects something other than 'most recently used'. When this happens, NVDA's focus object will move to the element with system focus on it because emoji edit field becomes hidden (location is None), causing NVDA to queue gain focus event. To prevent this, check to make sure the naviator object is the same as the UIA element with element selected event raised so that NVDA will nkow that a different part of the emoji panel is showing instead of redefining focus object (from NVDA's viewpoint). This allows NVDA to announce emoji panel content when categories such as symbls are selected.