Install tasks: removed unused supported releaes list. Fixes #92.
Now that minimum Windows version flag is in use, there is no need to loop through supported releases list. The code now resembles onInstall function of old but with edits to detect supported Windows 11 releases. In some ways, it returns instal tasks to 2021 and earlier but with experiences from subsequent years added.