Modern keyboard: remove the app module.
Modern keyboard app module provided some of the most noticeable features of this add-on. Now that all of its workarounds are included in NVDA (as of 2024.3), there is no need to keep it around unless workarounds are needed in emoji panel/clipboard history/dictation/voice typing/suggested actions/IME support in the future. Specificlaly, overlay class for navigation menu item, as well as event handlers for element selected, gain focus, and focus entered are no longer needed.