WinApps/UIAHandler/Ex: do not give up just becasue native window handle isn't fond.
Do not give up notification event handling if native window handle is 'absent ' - it can be found in runtime ID array (tuple) but there are times when that fails (tuple is empty). Therefore, return early if runtime ID cannot be retrieved, otherwise proceed to the rest of the event handler code.