From wiki to main repo: transfer add-on internals documentation from add-on wiki to the main code repository.
While the add-on internals document may belong to docs collection, the document is intended for developers wishing to learn more about the history and inner workings of the add-on. Despite reduced scope of the add-on, it is handy to describe how the add-on came to be and its workings to serve sa a record for future developers.