Copilot (WebView2 based): add app module for Copilot native app.
Copilot the progressive web app became a native app powered by WebView2 in late 2024. The ap module is needed to handle live region change event (used by Copilot to offer its responses) with a workarond: the section name is empty but the child element is Copilot response text. Therefore, tell NVDA to announce the (first) child element. Aso, copyright is set for 2025 as Copilot native app module is destined for add-on 25.01.