System settings: deprecate live region change event altogether as it is no loner raised in Windows 10/Windows UPdate screen.
As of January 2025 optional update (KB5050081/19045.5440), Settings app/Windows Update screen no longer raises live region change event. The issue with focus change with updates list while updates are downloading is not resolved although Microsoft was informed. Since this issue also does not happen with Narrator, mark live region change event as deprecated and remove it (and the app module) in a future add-on release.