App modules: add an empty __init__.py file.
With Copilot WebView2 app module removal, there is no longer any app modules in the add-on package. However, Git does not like empty folder commits, so add a dummy (__init__.py) file that will be present in the code repo only (the purpose of the dummy file is to become basis for an app module should the need arise).