WinApps: restore enhanced UIA notification event handler (Windows 11 Voice Access support).
In July 2024, work was done to support Windows 11 Voice Access notifications from everywhere by enhancing and patching UIA notification event handler. The work was removed in late July as the result was not generalizable (there were runtime Id's with three items instead of four). In May 2025, a community discussion on window state (restored/maximized) announcement took place in NVDA users list, prompting the July 2024 work to be revisited and thus restored with this commit. Apart from removing Windows version check in the lgoal plugin, the code is identical to July 2024 work, with enhancements incoming.