WinApps (global plugin): remove WinUI3 window class name addition again (resolved in NVDA 2025.2)
Originally planned to be removed in 25.08 but delayed to assess its impact. In some ways, WinUI3 is the final anchor to Windows 10 apartfrom install tasks message. As Windows 10 support is winding down, remove WinUI3 mouse/touch navigation support so the add-on can focus on Windows 11 support for the remainder of 2025.